+++ src/javasources/KTool/src/org/kframework/krun/api/TestGenResults.java
@@ -4,7 +4,7 @@ package org.kframework.krun.api;
 import java.util.List;
 import java.util.Map;
 
-import org.kframework.backend.unparser.UnparserFilterNew;
+import org.kframework.backend.unparser.UnparserFilter;
 import org.kframework.kil.Term;
 import org.kframework.kil.loader.Context;
 
@@ -36,7 +36,7 @@ public class TestGenResults {
             // TODO(YilongL): how to set state id?
             sb.append("\n\nTest case " + n /*+ ", State " + testGenResult.getState().getStateId()*/ + ":");
 
-            UnparserFilterNew t = new UnparserFilterNew(true, context.colorOptions.color(),
+            UnparserFilter t = new UnparserFilter(true, context.colorOptions.color(),
                     context.krunOptions.output, false, context);
             Term concretePgm = KRunState.concretize(testGenResult.getGeneratedProgram(), context);
             t.visitNode(concretePgm);
@@ -46,7 +46,7 @@ public class TestGenResults {
             Map<String, Term> substitution = testGenResult.getSubstitution();
 
             if (isDefaultPattern()) {
-                UnparserFilterNew unparser = new UnparserFilterNew(true, context.colorOptions.color(),
+                UnparserFilter unparser = new UnparserFilter(true, context.colorOptions.color(),
                         context.krunOptions.output, false, context);
                 unparser.visitNode(substitution.get("B:Bag"));
                 sb.append("\n" + unparser.getResult());
@@ -54,7 +54,7 @@ public class TestGenResults {
                 boolean empty = true;
 
                 for (String variable : substitution.keySet()) {
-                    UnparserFilterNew unparser = new UnparserFilterNew(true, context.colorOptions.color(),
+                    UnparserFilter unparser = new UnparserFilter(true, context.colorOptions.color(),
                             context.krunOptions.output, false, context);
                     sb.append("\n" + variable + " -->");
                     unparser.visitNode(substitution.get(variable));
