+++ src/javasources/KTool/test/org/kframework/backend/unparser/UnparserFilterNewTest.java
@@ -30,7 +30,7 @@ public class UnparserFilterNewTest {
      */
     @Test
     public void testTrailingUnderscores() {
-        UnparserFilterNew v = new UnparserFilterNew(context);
+        UnparserFilter v = new UnparserFilter(context);
         KApp t = KApp.of(KLabelConstant.of("'__", context), Variable.getFreshVar(Sort.K), Variable.getFreshVar(Sort.K));
         v.visit(t, null);
         v.getResult();
